% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{article}
\alias{article}
\title{Output format for INFORMS journal submissions}
\usage{
article(...)
}
\arguments{
\item{\dots}{Arguments passed to \code{\link[bookdown]{pdf_document2}}.}
}
\value{
An R Markdown output format object.
}
\description{
Each function is a wrapper for \code{\link[bookdown]{pdf_document2}} to
produce documents in INFORMS style.
}
\author{
Rob J Hyndman
}
